+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - [feat (#4): [syntax/parser, typechecker, codegen] return statements](https://github.com/amp-lang/amp/issues/4)
 - [feat (#8): [syntax/scanner, syntax/parser, typechecker, codegen] implement slice types](https://github.com/amp-lang/amp/issues/8)
 - [feat (#9): [syntax/scanner, syntax/parser, typechecker, codegen] implement variable declarations](https://github.com/amp-lang/amp/issues/9)
+- [feat (#11): [typechecker, codegen] implement function arguments](https://github.com/amp-lang/amp/issues/11)
 
 ### Compiler
 - feat: Basic stopwatch benchmark for Amp compile times.